agent_requirements.txt
duckduckgo-search
googlesearch-python
trafilatura
sqlalchemy
psycopg2-binary
requests
python-dateutil
